Refactor HomeScreenBasketAdapter
Renamed to HomeScreenAdapter
